Switch unit tests to pytest
