+++ pytest.ini
@@ -1,0 +1,12 @@
+[pytest]
+markers=
+    artifacts: tests relating to artifact monitoring
+    run: tests relating to simvue runs
+    offline: tests for systems with no server access
+    executor: tests of the simvue client executor
+    folder: tests of folder creation/edit
+    metadata: tests of metadata creation/edit
+    events: tests for events handling
+    metrics: tests for metrics handling
+    tags: tests for tag handling
+    unit: unit tests
