--- tests/unit/test_numpy_array_serialization.py
@@ -1,13 +1,0 @@
-from simvue.serialization import Serializer, Deserializer
-import numpy as np
-
-def test_numpy_array_serialization():
-    """
-    Check that a numpy array can be serialized then deserialized successfully
-    """
-    array = np.array([1, 2, 3, 4, 5])
-
-    serialized, mime_type = Serializer().serialize(array)
-    array_out = Deserializer().deserialize(serialized, mime_type)
-
-    assert (array == array_out).all()
