--- tests/unit/test_pytorch_tensor_serialization.py
@@ -1,14 +1,0 @@
-import torch
-from simvue.serialization import Serializer, Deserializer
-
-def test_pytorch_tensor_serialization():
-    """
-    Check that a PyTorch tensor can be serialized then deserialized successfully
-    """
-    torch.manual_seed(1724)
-    array = torch.rand(2, 3)
-
-    serialized, mime_type = Serializer().serialize(array)
-    array_out = Deserializer().deserialize(serialized, mime_type)
-
-    assert (array == array_out).all()
